requirements.txt
streamlit
python-dotenv
langchain
azure-identity
langchain_community
streamlit_extras